refactor: create an OpenAI client
