+++ usecases/modulecomponents/batch/tokens.go
@@ -9,7 +9,7 @@
 //  CONTACT: hello@weaviate.io
 //
 
-package clients
+package batch
 
 import (
 	"strings"
