+++ usecases/modulecomponents/batch/tokens_test.go
@@ -9,7 +9,7 @@
 //  CONTACT: hello@weaviate.io
 //
 
-package clients
+package batch
 
 import (
 	"testing"
